+++ ext/polars/src/lazyframe/mod.rs
@@ -1,8 +1,10 @@
+mod exitable;
 mod general;
 mod optflags;
 mod serde;
 mod sink;
 
+pub use exitable::RbInProcessQuery;
 use magnus::{TryConvert, Value};
 use polars::prelude::{Engine, LazyFrame, OptFlags};
 pub use sink::SinkTarget;